// src/Config/typescript/enums.ts
export enum Status {
  FOOD_MISSING = "FOOD_MISSING", //
  WATER_MISSING = "WATER_MISSING",
  FW_MISSING = "FW_MISSING", //
  OK = "OK", //
  NO_OK = "NO_OK", //
}

export enum StatusValue {
  FOOD_UPDATE_FILL = "FOOD_UPDATE_FILL", //
  WATER_UPDATE_FILL = "WATER_UPDATE_FILL", //
  FW_UPDATE_FILL = "FW_UPDATE_FILL", //
  FOOD_UPDATE_NO_FILL = "FOOD_UPDATE_NO_FILL", //
  WATER_UPDATE_NO_FILL = "WATER_UPDATE_NO_FILL", //
  FW_UPDATE_NO_FILL = "FW_UPDATE_NO_FILL", //
  PROBLEM = "PROBLEM",
  FIX = "FIX,", //
}